+++ src/components/AuditOrderList.js
@@ -1,0 +1,12 @@
+import AuditOrder from "./AuditOrder";
+
+const AuditOrderList = (props) => {
+  const component = props.orders?.map((order) => (
+    <AuditOrder key={`order-${order.id}`} order={order} />
+  ));
+  return (
+    <tbody className="screen__table-rows audit-order__list">{component}</tbody>
+  );
+};
+
+export default AuditOrderList;
